feat: all shemas in one module
