--- trainer_service/app/api/schemas/__init__.py
@@ -1,1 +1,0 @@
-pass
