+++ trainer_service/app/api/schemes/__init__.py
@@ -1,0 +1,5 @@
+from .metrices import MetricesParamCollections, MetricesParams
+from .tasker import Task, TaskParams, TaskStatus
+from .trainer import TrainerParams, LossConfig, ShedulerConfig, OptimizerConfig
+from .ml_models import ModelParams
+from .data import DataLoaderParams
