+++ trainer_service/app/api/schemes/data.py
