+++ trainer_service/app/api/schemes/metrices.py
