+++ trainer_service/app/core/datas/__init__.py
@@ -1,4 +1,3 @@
 from .loader import create_dataloaders
-from .shemas import DataLoaderParams
 
-__all__ = ['create_dataloaders', 'DataLoaderParams']
+__all__ = ['create_dataloaders']
