+++ trainer_service/app/core/models/__init__.py
@@ -1,4 +1,3 @@
 from .factory import get_model, get_models_type_name
-from .shemas import ModelParams
 
-__all__ = ['ModelParams', 'get_model', 'get_models_type_name']
+__all__ = ['get_model', 'get_models_type_name']
