+++ trainer_service/app/core/trainer/train_model.py
@@ -9,7 +9,7 @@ import asyncio
 
 from app.service.metrices import MetricesClient
 from app.service.tasker import TaskerClient
-from app.service.tasker.shemas import TrainerParams, LossConfig, OptimizerConfig, ShedulerConfig
+from app.api.schemes import TrainerParams, LossConfig, OptimizerConfig, ShedulerConfig
 from app.logs import get_logger
