+++ trainer_service/app/service/metrices/collection.py
@@ -6,7 +6,7 @@ from torchmetrics.classification import (
 )
 from torch import device
 
-from .shemas import MetricesParams, MetricesParamCollections
+from app.api.schemes import MetricesParams, MetricesParamCollections
 from app.logs import get_logger
 
 logger = get_logger(__name__)
