+++ trainer_service/app/service/tasker/__init__.py
@@ -1,2 +1,1 @@
 from .tasker import tasker_service, TaskerClient
-from .shemas import TaskStatus, TaskParams
